vinow-backend/requirements.txt
# FastAPI + Pydantic v2
fastapi==0.111.0
uvicorn==0.29.0

# Pydantic v2 系列
pydantic==2.7.1
pydantic-core==2.18.2
pydantic-settings==2.2.1

# Networking (必须用 0.24.1，否则冲突)
httpx==0.24.1
requests==2.32.2

# Supabase（兼容 httpx==0.24.1）
supabase==2.4.0

# Auth
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Tools
python-dotenv==1.0.1
pytest==8.2.0
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0

 # 更新 requirements.txt
# 新增依赖
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dateutil==2.8.2 # 更新 requirements.txt


内容模块

# FastAPI框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库和ORM
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Supabase客户端
supabase==2.3.1
postgrest==0.10.7
realtime==0.2.1

# Redis客户端
redis==5.0.1
hiredis==2.2.3

# 安全认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6

# 数据处理和验证
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# 文件处理
python-magic==0.4.27
aiofiles==23.2.1
pillow==10.1.0

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 异步任务
celery==5.3.4
redis==5.0.1

# 环境变量管理
python-dotenv==1.0.0

# 日志记录
structlog==23.2.0
python-json-logger==2.0.7

# 监控和性能
sentry-sdk==1.35.0
prometheus-client==0.19.0
psutil==5.9.6

# 测试框架
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# 代码质量和格式化
black==23.11.0
isort==5.13.2
flake8==6.1.0
mypy==1.7.1

# 开发工具
ipython==8.17.2
jupyter==1.0.0

# 其他工具
python-dateutil==2.8.2
pytz==2023.3
tzdata==2023.3

# FastAPI 核心
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
supabase==2.3.1
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.0

# 数据处理
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# 安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==41.0.8

# 文件处理
aiofiles==23.2.1
pillow==10.1.0
python-magic==0.4.27

# 搜索
elasticsearch==8.11.0

# 缓存
redis==5.0.1

# 消息队列
celery==5.3.4
redis==5.0.1

# HTTP 客户端
httpx==0.25.2
aiohttp==3.9.1

# 监控和日志
structlog==23.2.0
sentry-sdk==1.35.0
prometheus-client==0.18.0

# 工具类
pytz==2023.3
python-dateutil==2.8.2
jsonschema==4.19.2

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
